borg: make sure spell information is retrieved before borg_notice
